+++ src/igor/igor/resources/mesh/iMeshBuilderUtils.h
@@ -91,15 +91,17 @@ namespace igor
         */
         void IGOR_API addRing(iMeshBuilder &meshBuilder, float32 innerRadius, float32 outerRadius, uint32 segments = 16);
 
-        /*! adds a rectangle in XZ plane
+        /*! adds a plane to the given mesh builder instance
+
+        the plane is centered at the origin facing up
 
         \param meshBuilder the given mesh builder instance
-        \param x x position of plane
-        \param z z position of plane
-        \param sizeX size in x axis
-        \param sizeZ size in z axis
+        \param width the width along the x axis
+        \param depth the depth along the z axis
+        \param segmentsX the segment count along the x axis
+        \param segmentsZ the segment count along the z axis
         */
-        void IGOR_API addRectangle(iMeshBuilder &meshBuilder, float32 x, float32 z, float32 sizeX, float32 sizeZ);
+        void IGOR_API addPlane(iMeshBuilder &meshBuilder, float32 width, float32 depth, uint32 segmentsX = 16, uint32 segmentsZ = 16);
 
         /*! adds a mesh to the mesh builder